requirements.txt
paho-mqtt
google-cloud-speech
pyaudio